+++ src/KKdLib/database/texture.h
@@ -12,7 +12,7 @@
 class texture_info {
 public:
     std::string name;
-    uint64_t name_hash;
+    uint32_t name_hash;
     uint32_t id;
 
     texture_info();
